requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
pymongo==4.6.1
python-dotenv==1.0.0
passlib==1.7.4
requests==2.31.0
fpdf==1.7.2
gunicorn==21.2.0
eventlet==0.33.3
python-socketio==5.10.0
dnspython==2.4.2
python-engineio==4.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
urllib3==2.0.4
six==1.16.0
bidict==0.22.1
gevent==23.9.1
greenlet==2.0.2
zope.event==5.0
zope.interface==6.0 